updated animated sun components, added animated sun to the initial step
